requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
codecov==2.0.15
coverage==4.4.1
django-model-utils==3.1.2
idna==2.7                 # via requests
mock==2.0.0
pbr==4.2.0                # via mock
pytz==2018.5              # via django
requests==2.19.1          # via codecov
six==1.11.0               # via mock
urllib3==1.23             # via requests
